% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amova_missing.R
\name{amova_missing}
\alias{amova_missing}
\title{AMOVA tailored for genomic data with missing values (e.g. RADseq)}
\usage{
amova_missing(data, strata = NULL, ci = TRUE, quantiles.ci = c(0.025,
  0.975), method.ci = "percentile", iteration.ci = 10000, permute.ind = 0,
  subsample = NULL, iteration.subsample = NULL, mc = TRUE,
  mc.iterations = 100, mc.holdout = 0.5, digits = 9,
  parallel.core = parallel::detectCores() - 1, ...)
}
\arguments{
\item{data}{A file in the working directory or object in the global environment 
in wide or long (tidy) formats. See details for more info.

\emph{How to get a tidy data frame ?}
\href{https://github.com/thierrygosselin/stackr}{stackr} 
\code{\link[stackr]{tidy_genomic_data}} can transform 6 genomic data formats 
in a tidy data frame. You can also use this function to filter your dataset using
whitelist of markers, blacklist of individuals and genotypes.
Imputation of missing values is also offered with the function.}

\item{strata}{(optional for data frame and PLINK files, 
required for VCF and haplotypes files) A tab delimited file with 2 columns with header:
\code{INDIVIDUALS} and \code{STRATA}. With a 
data frame of genotypes the strata is the INDIVIDUALS and POP_ID columns, with
PLINK files, the \code{tfam} first 2 columns are used. 
If a \code{strata} file is specified, the strata file will have
precedence. The \code{STRATA} column can be any hierarchical grouping. 
To create a strata file see \code{\link[stackr]{individuals2strata}}.
If you have already run 
\href{http://catchenlab.life.illinois.edu/stacks/}{stacks} on your data, 
the strata file is similar to a stacks `population map file`, make sure you 
have the required column names  (\code{INDIVIDUALS} and \code{STRATA}).
Default: \code{strata = NULL}.}

\item{ci}{(logical, optional) Compute bootstrapped confidence intervals. 
Default: \code{ci = FALSE}.}

\item{quantiles.ci}{(double, optional) 
The quantiles for the bootstrapped confidence intervals. 
Default: \code{quantiles.ci = c(0.025,0.975)}.}

\item{method.ci}{(character, optional) With \code{method.ci = "percentile"},
the bootstrap confidence intervals are generated with the percentile interval
method. Upcomming: \code{method.ci = "bca"}, will allow 
the bootstrap confidence intervals to be generated with the adjusted bootstrap percentile interval (BCa) 
method (Efron, 1987)
provided by the package \href{https://CRAN.R-project.org/package=wBoot}{wBoot}.
Default: \code{method.ci = "percentile"}.}

\item{iteration.ci}{(integer, optional) The number of iterations for
the boostraps (resampling with replacement of markers).
Default: \code{iteration.ci = 10000}.}

\item{permute.ind}{(Integer) This argument allows to test 
sigma2a variance component of φst and get p.value by markers. 
Population labels are simply reordered.
Default: \code{permute.ind = 0} for no permutation}

\item{subsample}{(Integer) Default is no subsampling, \code{subsample = NULL}.
Using \code{subsample = "min"} will use the minimum sample size found in 
the data. 
With \code{subsample = 36}, 36 individuals in each populations are chosen
randomly to represent the dataset. If the \code{subsample} value is higher
than the minimum sample size, the minimum sample size found in the data will
be used as default.}

\item{iteration.subsample}{(Integer) need description here}

\item{mc}{(logical) With \code{mc = TRUE} the p.value and 
standard deviations of \emph{F}-statistics are computed using repeated random
sampling with a Monte Carlo procedure.
Default: \code{mc = FALSE}.}

\item{mc.iterations}{(integer) The number of random training-holdout data 
used for the Monte Carlo procedure.
Default: \code{mc.iterations = 100}.}

\item{mc.holdout}{(numeric, double) The proportion of holdout samples
used for the Monte Carlo procedure.
Default: \code{mc.holdout = 0.5}.}

\item{digits}{(optional, integer) The number of decimal places to be used in 
results.
Default: \code{digits = 9}.}

\item{parallel.core}{(optional) The number of core for parallel computation.
Default: \code{detectCores()-1}.}

\item{...}{other parameters passed to the function.}
}
\value{
With confidence intervals computed, the function returns a list with 
3 objects: per.locus = locus, global = global, global.boot.iterations = global.boot.iterations
\itemize{
 \item \code{$per.locus}: φst and φ'st per locus,
 \item \code{$global}: global φst and φ'st with confidence intervals,
 \item \code{$global.boot.iterations}: the boostrap φst and φ'st results
 along the random seed used during resampling.
}
}
\description{
\strong{amova_missing} computes standardized measures of population 
differentiation \emph{F}-statistics using the Analysis of Molecular Variance 
(AMOVA). The framework also uses the malhalanobis distance metric with 
the Choleski decomposition on the covariance matrix.
The speed is provided by the \pkg{tidyverse}, \pkg{multidplyr} and 
\pkg{parallel} packages.

\strong{Features:}

 \itemize{
 \item \strong{\emph{F}-statistics calculated:} correspond to those defined
 by Weir & Cockerham (1984) (φst, Fis, Fct) along the standardized versions 
 (φ'st) presented in Hedrick (2005) and Meirmans (2006).
 \item \strong{Confidence intervals:} based on bootstrap of markers 
 (sample WITH replacement),
 \item \strong{Monte Carlo:} to generate p.value and standard deviation of 
 \emph{F}-statistics, 
 \item \strong{Individual subsampling:} aim to reduce sampling unevenness
 between populations (sample WITHOUT replacement),
 \item \strong{Missing values: } allowed and used! As long as 
 the covariance matrix can be calculated. 
 \item \strong{Imputation: } Map-independent imputation of missing genotype 
 or alleles using Random Forest or the most frequent category is also 
 available to test the impact of missing data on AMOVA. The option is 
 provided through the use of \code{\link[stackr]{tidy_genomic_data}}.
 \item \strong{Hierarchy:} only population level analysis. 
 Future update will allow more hierarchy level.
 \item \strong{Ploidy:} currently build for diploid, but future update will 
 support more ploidy options.
}
}
\details{
\strong{Input data:}
 
To discriminate the long from the wide format, 
the function \pkg{stackr} \code{\link[stackr]{read_long_tidy_wide}} searches 
for \code{MARKERS or LOCUS} in column names (TRUE = long format).
The data frame is tab delimitted.
\strong{Wide format:}
The wide format cannot store metadata info.
The wide format starts with these 2 id columns: 
\code{INDIVIDUALS}, \code{POP_ID} (that refers to any grouping of individuals), 
the remaining columns are the markers in separate columns storing genotypes.

\strong{Long/Tidy format:}
The long format is considered to be a tidy data frame and can store metadata info. 
(e.g. from a VCF see \pkg{stackr} \code{\link{tidy_genomic_data}}). A minimum of 4 columns
are required in the long format: \code{INDIVIDUALS}, \code{POP_ID}, 
\code{MARKERS or LOCUS} and \code{GENOTYPE or GT}. The rest are considered metata info.

\strong{2 genotypes formats are available:}
6 characters no separator: e.g. \code{001002 of 111333} (for heterozygote individual).
6 characters WITH separator: e.g. \code{001/002 of 111/333} (for heterozygote individual).
The separator can be any of these: \code{"/", ":", "_", "-", "."}.

\emph{How to get a tidy data frame ?}
\pkg{stackr} \code{\link{tidy_genomic_data}} can transform 6 genomic data formats 
in a tidy data frame.
}
\note{
\strong{Negative φst} are technical artifact of the computation 
(see Roesti el al. 2012) and are automatically replaced with zero inside 
this function.

\strong{Roadmap of future developments: } 
     \itemize{
     \item \strong{Hierarchy: } More hierarchy
     \item \strong{Ploidy:} More ploidy
     \item \strong{markers statistics: } p.value/confidence intervals statistics
     for individual markers present in a whitelist (by resampling their variance component).
      \item \strong{Genotype Likelihoods: } Instead of allele count...
      }
}
\examples{
\dontrun{
wombat.amova <- amova_missing(
data = "wombat.filtered.tidy.tsv", 
ci = TRUE, 
quantiles.ci = c(0.025,0.975),
subsample = 30,
iteration.ci = 100, 
mc = TRUE, 
mc.iterations = 10000, 
mc.holdout = 0.5,
parallel.core = 8,
)
To get the global estimates:
wombat.amova$global
To get the per locus estimates:
wombat.amova$per.locus
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\references{
Excoffier L, Smouse PE, Quattro JM. 
Analysis of molecular variance inferred from metric distances among 
DNA haplotypes: application to human mitochondrial DNA restriction data. 
Genetics. 1992;131: 479-491.

Meirmans PG, Van Tienderen PH (2004) genotype and genodive: 
two programs for the analysis of genetic diversity of asexual organisms. 
Molecular Ecology Notes, 4, 792-794.

Michalakis Y, Excoffier L. 
A generic estimation of population 
subdivision using distances between alleles with special reference for 
microsatellite loci. 
Genetics. 1996;142: 1061-1064.

Weir BS, Cockerham CC (1984) Estimating F-Statistics for the
Analysis of Population Structure. 
Evolution, 38, 1358-1370.

Roesti M, Salzburger W, Berner D. (2012)
Uninformative polymorphisms bias genome scans for signatures of selection. 
BMC Evol Biol., 12:94. doi:10.1111/j.1365-294X.2012.05509.x

Meirmans PG. 2006. 
Using the amova framework to estimate a standardized genetic differentiation
measure.
Evolution. 60: 2399-2402. 
doi:10.1111/j.0014-3820.2006.tb01874.x

Hedrick PW (2005) 
A standardized genetic differentiation measure. 
Evolution, 59, 1633-1638.

Efron, B. (1987) Better bootstrap confidence intervals (with Discussion). 
Journal of the American Statistical Association, 82, 171-200.
}
\seealso{
\href{http://www.bentleydrummer.nl/software/software/GenoDive.html}{GenoDive}

\code{\link[stackr]{tidy_genomic_data}} to transform numerous genomic data 
format in tidy data frames.
}

