% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_WC84.R
\name{assigner_fst_stats}
\alias{assigner_fst_stats}
\title{assigner_fst_stats}
\usage{
assigner_fst_stats(
  x,
  v,
  group.by = NULL,
  outliers = FALSE,
  overall = FALSE,
  keep.groups = "drop",
  digits = NULL
)
}
\description{
Generate useful stats
}
\keyword{internal}
