% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assigner_function_header}
\alias{assigner_function_header}
\title{assigner_function_header}
\usage{
assigner_function_header(f.name = NULL, start = TRUE, verbose = TRUE)
}
\description{
Generate function header
}
\keyword{internal}
