% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assigner_toc}
\alias{assigner_toc}
\title{assigner_toc}
\usage{
assigner_toc(
  timing,
  end.message = "Computation time, overall:",
  verbose = TRUE
)
}
\description{
assigner tictoc function
}
\keyword{internal}
