% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_mixture.R
\name{assignment_gsi_mix}
\alias{assignment_gsi_mix}
\title{assignment with gsi_sim}
\usage{
assignment_gsi_mix(input, strata.df = NULL, select.markers = NULL,
  markers.names = NULL, missing.data = NULL, i = NULL, m = NULL,
  filename = NULL, directory.subsample = NULL,
  keep.gsi.files = FALSE, pop.labels = NULL,
  sampling.method = "random", ...)
}
\description{
assignment with gsi_sim
}
\keyword{internal}
