% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_assignment.R
\name{assignment_gsi_sim}
\alias{assignment_gsi_sim}
\title{Assignment with Eric C. Anderson's gsi_sim}
\usage{
assignment_gsi_sim(
  input = NULL,
  strata = NULL,
  select.markers = NULL,
  i = NULL,
  m = NULL,
  holdout = NULL,
  filename = NULL,
  i.folder = NULL,
  keep.gsi.files = FALSE,
  markers.sampling = "random",
  thl = 1
)
}
\description{
Run population assignment with
  \href{https://github.com/eriqande/gsi_sim}{gsi_sim}, developed by assigner
  co-author Eric C. Anderson. Eric also developed the original wrapper used
  by \pkg{assigner}.
}
\keyword{internal}
