% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_assignment.R
\name{assignment_random}
\alias{assignment_random}
\title{assignment_random}
\usage{
assignment_random(
  x,
  assignment.analysis = "gsi_sim",
  input = NULL,
  genind.object = NULL,
  strata = NULL,
  directory.subsample = NULL,
  keep.gsi.files = FALSE,
  markers.sampling = "random",
  subsample.id = NULL,
  filename = NULL,
  adegenet.n.rep = 30,
  adegenet.training = 0.9,
  holdout = NULL
)
}
\description{
assignment_random
}
\keyword{internal}
