% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_assignment.R
\name{assignment_subsamples}
\alias{assignment_subsamples}
\title{assignment_subsamples}
\usage{
assignment_subsamples(
  x,
  input = NULL,
  subsample = NULL,
  assignment.analysis = "gsi_sim",
  marker.number = NULL,
  markers.sampling = "random",
  iteration.method = 10,
  filename = "assignment_data.txt",
  directory = NULL,
  keep.gsi.files = FALSE,
  verbose = FALSE,
  parallel.core = parallel::detectCores() - 1,
  random.seed = NULL,
  base.filename = NULL,
  thl = 1,
  adegenet.dapc.opt = NULL,
  adegenet.n.rep = NULL,
  adegenet.training = NULL
)
}
\description{
assignment_subsamples
}
\keyword{internal}
