% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_WC84.R
\name{compute_fst}
\alias{compute_fst}
\title{compute_fst}
\usage{
compute_fst(x, ci = FALSE, iteration.ci = 100,
  quantiles.ci = c(0.025, 0.975), digits = 9, path.folder = NULL)
}
\description{
main function
}
\keyword{internal}
