% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_NEI87.R
\name{compute_fst_nei}
\alias{compute_fst_nei}
\title{compute_fst_nei}
\usage{
compute_fst_nei(x, ci = FALSE, iteration.ci = 100,
  quantiles.ci = c(0.025, 0.975), digits = 9)
}
\description{
fst function
}
\keyword{internal}
