% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_mean_range}
\alias{format_mean_range}
\title{Format mean and range of numeric vector as string}
\usage{
format_mean_range(
  x,
  meanx = TRUE,
  min.max = TRUE,
  digits = 3,
  scientific = FALSE
)
}
\arguments{
\item{x}{A numeric vector.}

\item{meanx}{Logical; whether to include the mean value in the output. Defaults to `TRUE`.}

\item{min.max}{Logical; whether to include the minimum and maximum values in the output. Defaults to `TRUE`.}

\item{digits}{Number of digits *after the decimal point*.}

\item{scientific}{Logical; whether to use scientific notation.}
}
\value{
A character string with formatted mean, min, and max values.
}
\description{
Produces a string like "0.123 [0.100 - 0.150]" for a numeric vector.
}
\keyword{internal}
