% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_NEI87.R
\name{fst_nei_subsample}
\alias{fst_nei_subsample}
\title{fst_nei_subsample}
\usage{
fst_nei_subsample(x, input, pop.levels = NULL, pop.labels = NULL,
  strata = NULL, holdout.samples = NULL, pairwise = FALSE, ci = FALSE,
  iteration.ci = 100, quantiles.ci = c(0.025, 0.975), digits = 9,
  subsample = NULL, parallel.core = parallel::detectCores() - 1,
  verbose = FALSE)
}
\description{
Function that link all with subsampling
}
\keyword{internal}
