% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsi_sim_binary_support.R
\name{gsi_sim_binary}
\alias{gsi_sim_binary}
\title{file path to be used in a call to gsi_sim.}
\usage{
gsi_sim_binary()
}
\description{
This version checks to make sure it is there and throws an
error with a suggestion of how to get it if it is not there.
}

