% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsi_sim_binary_support.R
\name{gsi_sim_binary_path}
\alias{gsi_sim_binary_path}
\title{return the path where gsi_sim should be in the R system paths}
\usage{
gsi_sim_binary_path()
}
\description{
return the path where gsi_sim should be in the R system paths
}
\keyword{internal}

