% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsi_sim_binary_support.R
\name{gsi_sim_exists}
\alias{gsi_sim_exists}
\title{return TRUE if gsi_sim exists where it should be}
\usage{
gsi_sim_exists()
}
\description{
return TRUE if gsi_sim exists where it should be
}
\keyword{internal}
