% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_subsamples.R
\name{import_subsamples}
\alias{import_subsamples}
\title{Import individual's assignment results of different subsample folder.}
\usage{
import_subsamples(dir.path, imputations)
}
\arguments{
\item{dir.path}{The path to the directory containing the subsample folders.}

\item{imputations}{(logical) Was the data imputed or not. 
Default: \code{imputations = FALSE}}
}
\value{
A data frame of all the individual's assignment, with iterations and subsample.
}
\description{
This function will import all the individual's assignment results 
of different subsample folder into R.
}
\examples{
\dontrun{
subsamples.data <- import_subsamples(
dir.path = "assignment_analysis_method_random_imputations_rf_populations",
imputations = TRUE
)
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}

