% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_mixture.R
\name{mixture_baseline_intersect}
\alias{mixture_baseline_intersect}
\title{mixture_baseline_intersect}
\usage{
mixture_baseline_intersect(mixture.id, inferred.baseline,
  input.mixture = NULL, input.baseline = NULL)
}
\description{
mixture_baseline_intersect
}
\keyword{internal}
