% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_WC84.R
\name{pairwise_fst}
\alias{pairwise_fst}
\title{pairwise_fst}
\usage{
pairwise_fst(
  x,
  pop.pairwise = NULL,
  ci = FALSE,
  iteration.ci = 100,
  quantiles.ci = c(0.025, 0.975),
  digits = 9,
  path.folder = NULL,
  ...
)
}
\description{
main function
}
\keyword{internal}
