% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_NEI87.R
\name{pairwise_fst_nei}
\alias{pairwise_fst_nei}
\title{pairwise_fst_nei}
\usage{
pairwise_fst_nei(
  list.pair,
  pop.pairwise = NULL,
  data.genotyped = NULL,
  ci = FALSE,
  iteration.ci = 100,
  quantiles.ci = c(0.025, 0.975),
  digits = 9
)
}
\description{
pairwise Fst function for nei
}
\keyword{internal}
