% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_WC84.R
\name{pairwise_fst_snprelate}
\alias{pairwise_fst_snprelate}
\title{pairwise_fst_snprelate}
\usage{
pairwise_fst_snprelate(pop.pairwise, data, strata, unique.markers.pop)
}
\description{
Pairwise Fst function with SNPRelate
}
\keyword{internal}
