% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{plot_assignment_dlr}
\alias{plot_assignment_dlr}
\title{Assignment plot of genotype likelihood.}
\arguments{
\item{data}{The output assignment file (home likelihood or
likelihood ratio statistics) from GENODIVE.}

\item{l.skip}{(integer) The number of lines to skip before the individuals info
in GenoDive assignment results (see Vignette).}

\item{number.individuals}{(integer) The number of individuals analysed.}

\item{number.pop}{(integer) The number of populations analysed.}

\item{pop.id.start}{(Optional) The start of your population id
in the name of your individual sample. Your individuals are identified 
in this form : SPECIES-POPULATION-MATURITY-YEAR-ID = CHI-QUE-ADU-2014-020,
then, \code{pop.id.start} = 5. If you didn't name your individuals
with the pop id in it, use the \code{strata} argument.}

\item{pop.id.end}{(Optional) The end of your population id
in the name of your individual sample. Your individuals are identified 
in this form : SPECIES-POPULATION-MATURITY-YEAR-ID = CHI-QUE-ADU-2014-020,
then, \code{pop.id.end} = 7. If you didn't name your individuals
with the pop id in it, use the \code{strata} argument.}

\item{strata}{(optional) A tab delimited file with 2 columns with header:
\code{INDIVIDUALS} and \code{STRATA}. Default: \code{strata = NULL}.
The \code{STRATA} column is used here as the populations id of your sample.}

\item{pop.levels}{A character string with your sampling sites or populations 
in the order you prefer (for tables and figure).}

\item{pop.labels}{(optional) A character string for your populations labels.
If you need to rename sampling sites in \code{pop.levels} or combined sites/pop
into a different names, here is the place.}

\item{POPA}{First population to compare.}

\item{POPB}{Second population to compare (with A).}

\item{dlr}{(optional) Character string with Dlr value.}

\item{x.dlr}{(optional) Position to the x-axis of the Dlr value.}

\item{y.dlr}{(optional) Position to the y-axis of the Dlr value.}

\item{fst}{(optional) Character string with Fst value.}

\item{x.fst}{(optional) Position to the x-axis of the Fst value.}

\item{y.fst}{(optional) Position to the y-axis of the Fst value.}

\item{filename}{(optional) Name of the figure written
in the working directory.}

\item{plot.width}{(optional) Width in cm of the figure.}

\item{plot.height}{(optional) height in cm of the figure.}

\item{plot.dpi}{(optional) Number of dpi for the figure (e.g 600).}
}
\value{
A list with the assignment table and the assignment plot.
}
\description{
Create a figure similar to Paetkau's et al. (2004) Fig 6.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\references{
Paetkau D, Slade R, Burden M, Estoup A (2004)
Genetic assignment methods for the direct, real-time estimation of migration
rate: a simulation-based exploration of accuracy and power.
Molecular Ecology, 13, 55-65.

Meirmans PG, Van Tienderen PH (2004) genotype and genodive: 
two programs for the analysis of genetic diversity of asexual organisms. 
Molecular Ecology Notes, 4, 792-794.
}

