% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dlr.R
\name{plot_dlr}
\alias{plot_dlr}
\title{Assignment plot of genotype likelihood (Dlr).}
\usage{
plot_dlr(data, POPA, POPB, dlr = NULL, x.dlr = NULL, y.dlr = NULL,
  fst = NULL, x.fst = NULL, y.fst = NULL, filename = NULL,
  plot.width = NULL, plot.height = NULL, plot.dpi = NULL)
}
\arguments{
\item{data}{The assigment object from 
\code{\link[assigner]{dlr}} output.}

\item{POPA}{First population to compare.}

\item{POPB}{Second population to compare (with A).}

\item{dlr}{(optional) Character string with Dlr value.}

\item{x.dlr}{(optional) Position to the x-axis of the Dlr value.}

\item{y.dlr}{(optional) Position to the y-axis of the Dlr value.}

\item{fst}{(optional) Character string with Fst value.}

\item{x.fst}{(optional) Position to the x-axis of the Fst value.}

\item{y.fst}{(optional) Position to the y-axis of the Fst value.}

\item{filename}{(optional) Name of the figure written
in the working directory.}

\item{plot.width}{(optional) Width in cm of the figure.}

\item{plot.height}{(optional) height in cm of the figure.}

\item{plot.dpi}{(optional) Number of dpi for the figure (e.g 600).}
}
\value{
A list with the assignment table and the assignment plot.
}
\description{
The function generate a figure similar to Paetkau's et al. (2004)
Fig 6.
}
\examples{
\dontrun{
# if you didn't run assigner::dlr before:
dlr.fig <- assigner::dlr(
data = "assignment.gdv", strata = "my.strata.tsv")$assignment \%>\%
assigner::plot_dlr(data = ., POPA = "mypop1", POPB = "mypop2")

# to get the plot:
dlr$assignment.plot

# If you already have the results of assigner::dlr:
assigner::plot_dlr(data = results, POPA = "mypop1", POPB = "mypop2")$assignment.plot
}
}
\references{
Paetkau D, Slade R, Burden M, Estoup A (2004)
Genetic assignment methods for the direct, real-time estimation of migration
rate: a simulation-based exploration of accuracy and power.
Molecular Ecology, 13, 55-65.

Meirmans PG, Van Tienderen PH (2004) genotype and genodive: 
two programs for the analysis of genetic diversity of asexual organisms. 
Molecular Ecology Notes, 4, 792-794.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
