% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_WC84.R
\name{prep_strata}
\alias{prep_strata}
\title{Prepare stratification data}
\usage{
prep_strata(
  data,
  strata,
  pop.levels = NULL,
  blacklist.id = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{A tibble containing genetic data.}

\item{strata}{Path to the strata file or a tibble containing strata information.}

\item{pop.levels}{(optional) A character vector specifying the population levels to use for factor levels of `STRATA`. Default: \code{NULL}.}

\item{blacklist.id}{(optional) A vector of individual IDs to blacklist. Default: \code{NULL}.}

\item{verbose}{(logical) Show progress messages. Default: \code{FALSE}.}
}
\value{
A tibble with updated `STRATA` column and joined strata info.
}
\description{
Read and join strata information to the main dataset, standardising population levels.
}
\keyword{internal}
