% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_mixture.R
\name{subsampling_baseline}
\alias{subsampling_baseline}
\title{subsampling baseline data}
\usage{
subsampling_baseline(iteration.subsample = 1, ind.pop.df = NULL,
  subsample = NULL, random.seed = NULL, ...)
}
\description{
subsampling baseline data
}
\keyword{internal}
