% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subsampling_data}
\alias{subsampling_data}
\title{subsampling data}
\usage{
subsampling_data(iteration.subsample = 1, strata = NULL,
  subsample = NULL, random.seed = NULL)
}
\description{
subsampling data
}
\keyword{internal}
