% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils_formatC}
\alias{utils_formatC}
\title{Format numeric using formatC and dynamic arguments}
\usage{
utils_formatC(x, fmt.args)
}
\arguments{
\item{x}{A numeric scalar.}

\item{fmt.args}{A named list of arguments to pass to base::formatC.}
}
\value{
A character string formatted by formatC.
}
\description{
Format numeric using formatC and dynamic arguments
}
\keyword{internal}
