% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils_formatC_vec}
\alias{utils_formatC_vec}
\title{Format a numeric vector using formatC and flexible arguments}
\usage{
utils_formatC_vec(x, fmt.args)
}
\arguments{
\item{x}{A numeric vector.}

\item{fmt.args}{A named list of arguments to pass to base::formatC.}
}
\value{
A character vector with formatted values.
}
\description{
Format a numeric vector using formatC and flexible arguments
}
\keyword{internal}
