% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gsi_sim.R
\name{write_gsi_sim}
\alias{write_gsi_sim}
\title{Used internally in assigner to write a gsi_sim file}
\usage{
write_gsi_sim(data, markers.names = NULL, directory.subsample = NULL,
  filename = "gsi_sim.unname.txt", i = NULL, m = NULL, ...)
}
\arguments{
\item{data}{A file or object in the global environment containing at least 
these 2 id columns: 
\code{INDIVIDUALS}, \code{POP_ID} (that refers to any grouping of individuals.), 
the remaining columns are the markers.}

\item{markers.names}{(character) Name of the markers.}

\item{directory.subsample}{The name of the subsample directory.}

\item{filename}{The name of the file written to the working directory.}

\item{i}{Iterations.}

\item{m}{Markers number used.}

\item{...}{other parameters passed to the function.}
}
\value{
A gsi_sim file is saved to the working directory.
}
\description{
Write a gsi_sim file
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}

