% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assigner_dots.R
\name{assigner_dots}
\alias{assigner_dots}
\title{assigner_dots}
\usage{
assigner_dots(
  func.name = as.list(sys.call())[[1]],
  fd = NULL,
  args.list = NULL,
  dotslist = NULL,
  keepers = c("subsample.markers.stats", "force.stats", "id.stats", "subsample",
    "filter.reproducibility", "filter.individuals.missing",
    "filter.individuals.heterozygosity", "filter.individuals.coverage.total",
    "filter.individuals.coverage.median", "filter.individuals.coverage.iqr",
    "filter.common.markers", "filter.monomorphic", "filter.mac", "filter.coverage", "dp",
    "filter.genotyping", "filter.snp.position.read", "filter.snp.number",
    "filter.short.ld", "filter.long.ld", "long.ld.missing", "ld.method", "ld.figures",   
      "detect.mixed.genomes", "ind.heterozygosity.threshold", "detect.duplicate.genomes",
    "filter.hwe", "filter.strands", "random.seed", "path.folder", "filename",
    "parameters", "blacklist.genotypes", "erase.genotypes", "pop.levels", "pop.labels",
    "pop.select", "blacklist.id", "markers.info", "keep.allele.names", "keep.gds",
    "vcf.metadata", "vcf.stats", "wide", "whitelist.markers", "write.tidy",
    "missing.memory", "dart.sequence", "internal", "tidy.check", "tidy.vcf", "tidy.dart",
    "gt", "gt.bin", "gt.vcf",      "gt.vcf.nuc", "calibrate.alleles", "forking"),
  deprecated = NULL,
  verbose = TRUE
)
}
\arguments{
\item{func.name}{Default: \code{as.list(sys.call())[[1]]}.}

\item{fd}{(optional) Default: \code{rlang::fn_fmls_names()}.}

\item{args.list}{(optional) Default:\code{args.list = as.list(environment())}.}

\item{dotslist}{The dots dots dots arguments captured.
Default: \code{
dotslist = rlang::dots_list(..., .homonyms = "error", .check_assign = TRUE)}.}

\item{keepers}{(optional) The arguments that are used inside the function.
Default: \code{keepers = c(
"subsample.markers.stats", "subsample",
"filter.reproducibility", "filter.individuals.missing",
"filter.individuals.heterozygosity", "filter.individuals.coverage.total",
"filter.individuals.coverage.median", "filter.individuals.coverage.iqr",
"filter.common.markers", "filter.monomorphic", "filter.mac",
"filter.coverage", "filter.genotyping", "filter.snp.position.read",
"filter.snp.number", "filter.short.ld", "filter.long.ld", "long.ld.missing",
"ld.method", "detect.mixed.genomes", "ind.heterozygosity.threshold",
"detect.duplicate.genomes",
"filter.hwe", "filter.strands", "random.seed", "path.folder", "filename",
"blacklist.genotypes", "erase.genotypes",
"gt", "gt.bin", "gt.vcf", "gt.vcf.nuc",
"pop.levels", "pop.labels", "pop.select", "blacklist.id",
"markers.info", "keep.allele.names", "keep.gds",
"vcf.metadata", "vcf.stats", "id.stats", "dp",
"whitelist.markers",
"write.tidy",
"dart.sequence",
"missing.memory",
"internal",
"tidy.check",
"tidy.vcf", "tidy.dart",
"calibrate.alleles",
"forking"
)}.}

\item{deprecated}{(optional) assigner's deprecated arguments.
Default: \code{deprecated = NULL}.}

\item{verbose}{(logical) Function will output more details.}
}
\description{
Extract and assign the dots-dots-dots
}
\keyword{internal}
