% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_ngs.R
\name{assignment_adegenet}
\alias{assignment_adegenet}
\title{assignment with adegenet}
\usage{
assignment_adegenet(data = NULL, select.markers = NULL,
  adegenet.dapc.opt = "optim.a.score", adegenet.n.rep = 30,
  adegenet.training = 0.9, i = NULL, m = NULL, holdout = NULL,
  markers.sampling = "random", thl = 1, subsample.id = NULL,
  parallel.core = parallel::detectCores() - 1)
}
\description{
assignment with adegenet
}
\keyword{internal}
