% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_ngs.R
\name{assignment_function}
\alias{assignment_function}
\title{assignment_function}
\usage{
assignment_function(x, input = NULL, subsample = NULL,
  assignment.analysis = "gsi_sim", snp.ld = NULL, common.markers = TRUE,
  maf.thresholds = NULL, marker.number = NULL, pop.levels = NULL,
  pop.labels = NULL, sampling.method = "random", iteration.method = 10,
  filename = "assignment_data.txt", directory = NULL,
  keep.gsi.files = FALSE, imputation.method = NULL,
  impute.mixture = FALSE, hierarchical.levels = "populations",
  verbose = FALSE, parallel.core = parallel::detectCores() - 1,
  manage.all = NULL, random.seed = NULL, base.filename = NULL,
  res.list = NULL, thl = 1, adegenet.dapc.opt = NULL,
  adegenet.n.rep = NULL, adegenet.training = NULL)
}
\description{
assignment_function
}
\keyword{internal}
