% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_ngs.R
\name{assignment_gsi_sim}
\alias{assignment_gsi_sim}
\title{assignment with gsi_sim}
\usage{
assignment_gsi_sim(
  input = NULL,
  strata = NULL,
  select.markers = NULL,
  i = NULL,
  m = NULL,
  holdout = NULL,
  filename = NULL,
  i.folder = NULL,
  keep.gsi.files = FALSE,
  markers.sampling = "random",
  thl = 1
)
}
\description{
assignment with gsi_sim
}
\keyword{internal}
