% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_mixture.R, R/assignment_ngs.R
\name{assignment_gsi_sim}
\alias{assignment_gsi_sim}
\alias{assignment_gsi_sim}
\title{assignment with gsi_sim}
\usage{
assignment_gsi_sim(data = NULL, strata.df = NULL, select.markers = NULL,
  markers.names = NULL, missing.data = NULL, i = NULL, m = NULL,
  holdout = NULL, filename = NULL, directory.subsample = NULL,
  keep.gsi.files = FALSE, pop.labels = NULL, sampling.method = "random",
  thl = 1)

assignment_gsi_sim(data = NULL, strata.df = NULL, select.markers = NULL,
  markers.names = NULL, missing.data = NULL, i = NULL, m = NULL,
  holdout = NULL, filename = NULL, directory.subsample = NULL,
  keep.gsi.files = FALSE, pop.labels = NULL, sampling.method = "random",
  thl = 1)
}
\description{
assignment with gsi_sim

assignment with gsi_sim
}
\keyword{internal}
