% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_ngs.R
\name{assignment_marker_loop}
\alias{assignment_marker_loop}
\title{assignment_marker_loop}
\usage{
assignment_marker_loop(
  m,
  assignment.analysis = "gsi_sim",
  fst.ranked = NULL,
  i = NULL,
  input = NULL,
  genind.object = NULL,
  strata = NULL,
  markers.sampling = NULL,
  subsample.id = NULL,
  holdout = NULL,
  thl = 1,
  adegenet.dapc.opt = NULL,
  adegenet.n.rep = NULL,
  adegenet.training = NULL,
  i.folder = NULL,
  filename = NULL,
  keep.gsi.files = FALSE
)
}
\description{
assignment_marker_loop
}
\keyword{internal}
