% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_ngs.R
\name{assignment_ranking}
\alias{assignment_ranking}
\title{assignment_ranking}
\usage{
assignment_ranking(iterations.list = NULL, thl = NULL, input = NULL,
  pop.labels = NULL, input.imp = NULL, holdout.individuals.list = NULL,
  imputation.method = NULL, imputations.group = NULL,
  directory.subsample = NULL, marker.number = NULL,
  assignment.analysis = NULL, genind.object = NULL,
  genind.object.imp = NULL, strata.df = NULL, sampling.method = NULL,
  iteration.method = NULL, subsample.id = NULL, adegenet.dapc.opt = NULL,
  adegenet.n.rep = NULL, adegenet.training = NULL, filename = NULL,
  keep.gsi.files = NULL, parallel.core = parallel::detectCores() - 1, ...)
}
\description{
assignment_ranking
}
\keyword{internal}

