% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_ngs.R
\name{assignment_ranking}
\alias{assignment_ranking}
\title{assignment_ranking}
\usage{
assignment_ranking(iterations.list = NULL, thl = NULL, input = NULL,
  holdout.individuals = NULL, directory.subsample = NULL,
  marker.number = NULL, assignment.analysis = NULL,
  genind.object = NULL, strata = NULL, markers.sampling = NULL,
  subsample.id = NULL, adegenet.dapc.opt = NULL,
  adegenet.n.rep = NULL, adegenet.training = NULL, filename = NULL,
  keep.gsi.files = NULL)
}
\description{
assignment_ranking
}
\keyword{internal}
