% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlr.R
\name{dlr}
\alias{dlr}
\title{Genotype likelihood ratio distance (Dlr)}
\arguments{
\item{data}{The output assignment file (home likelihood or
likelihood ratio statistics) from GENODIVE.}

\item{l.skip}{(integer) The number of lines to skip before the individuals info
in GenoDive assignment results (see Vignette).}

\item{number.individuals}{(integer) The number of individuals analysed.}

\item{number.pop}{(integer) The number of populations analysed.}

\item{pop.id.start}{(Optional) The start of your population id
in the name of your individual sample. Your individuals are identified 
in this form : SPECIES-POPULATION-MATURITY-YEAR-ID = CHI-QUE-ADU-2014-020,
then, \code{pop.id.start} = 5. If you didn't name your individuals
with the pop id in it, use the \code{strata} argument.}

\item{pop.id.end}{(Optional) The end of your population id
in the name of your individual sample. Your individuals are identified 
in this form : SPECIES-POPULATION-MATURITY-YEAR-ID = CHI-QUE-ADU-2014-020,
then, \code{pop.id.end} = 7. If you didn't name your individuals
with the pop id in it, use the \code{strata} argument.}

\item{strata}{(optional) A tab delimited file with 2 columns with header:
\code{INDIVIDUALS} and \code{STRATA}. Default: \code{strata = NULL}.
The \code{STRATA} column is used here as the populations id of your sample.}

\item{pop.levels}{A character string with your sampling sites or populations 
in the order you prefer (for tables and figure).}

\item{pop.labels}{(optional) A character string for your populations labels.
If you need to rename sampling sites in \code{pop.levels} or combined sites/pop
into a different names, here is the place.}

\item{filename}{(optional) Name of the file prefix for
the matrix and the table written in the working directory.}
}
\value{
A list with 3 objects of class: table ($dlr.table), dist (a lower
diagonal matrix, $dlr.dist), data.frame (a mirrored matrix, $dlr.matrix).
}
\description{
The function computes Paetkau's et al. (1997) genotype likelihood
ratio distance (Dlr).
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\references{
Paetkau D, Slade R, Burden M, Estoup A (2004) 
Genetic assignment methods for the direct, real-time estimation of 
migration rate: a simulation-based exploration of accuracy and power. 
Molecular Ecology, 13, 55-65.

Paetkau D, Waits LP, Clarkson PL, Craighead L, Strobeck C (1997)
 An empirical evaluation of genetic distance statistics using microsatellite
  data from bear (Ursidae) populations. Genetics, 147, 1943-1957.

Meirmans PG, Van Tienderen PH (2004) genotype and genodive: 
two programs for the analysis of genetic diversity of asexual organisms. 
Molecular Ecology Notes, 4, 792-794.
}

