% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlr.R
\name{dlr}
\alias{dlr}
\title{Genotype likelihood ratio distance (Dlr)}
\usage{
dlr(data, l.skip, number.individuals, number.pop, pop.id.start = NULL,
  pop.id.end = NULL, pop.levels = NULL, pop.labels = NULL,
  strata = NULL, filename = NULL)
}
\arguments{
\item{data}{The output assignment file (home likelihood or
likelihood ratio statistics) from GENODIVE.}

\item{l.skip}{(integer) The number of lines to skip before the individuals info
in GenoDive assignment results (see Vignette).}

\item{number.individuals}{(integer) The number of individuals analysed.}

\item{number.pop}{(integer) The number of populations analysed.}

\item{pop.id.start}{(Optional) The start of your population id
in the name of your individual sample. Your individuals are identified 
in this form : SPECIES-POPULATION-MATURITY-YEAR-ID = CHI-QUE-ADU-2014-020,
then, \code{pop.id.start} = 5. If you didn't name your individuals
with the pop id in it, use the \code{strata} argument.}

\item{pop.id.end}{(Optional) The end of your population id
in the name of your individual sample. Your individuals are identified 
in this form : SPECIES-POPULATION-MATURITY-YEAR-ID = CHI-QUE-ADU-2014-020,
then, \code{pop.id.end} = 7. If you didn't name your individuals
with the pop id in it, use the \code{strata} argument.}

\item{pop.levels}{(optional, string) This refers to the levels in a factor. In this 
case, the id of the pop.
Use this argument to have the pop ordered your way instead of the default 
alphabetical or numerical order. e.g. \code{pop.levels = c("QUE", "ONT", "ALB")} 
instead of the default \code{pop.levels = c("ALB", "ONT", "QUE")}. 
Default: \code{pop.levels = NULL}.}

\item{pop.labels}{(optional, string) Use this argument to rename/relabel
your pop or combine your pop. e.g. To combine \code{"QUE"} and \code{"ONT"} 
into a new pop called \code{"NEW"}:
(1) First, define the levels for your pop with \code{pop.levels} argument: 
\code{pop.levels = c("QUE", "ONT", "ALB")}. 
(2) then, use \code{pop.labels} argument: 
\code{pop.levels = c("NEW", "NEW", "ALB")}.#' 
To rename \code{"QUE"} to \code{"TAS"}:
\code{pop.labels = c("TAS", "ONT", "ALB")}.
Default: \code{pop.labels = NULL}. If you find this too complicated, there is also the
\code{strata} argument that can do the same thing, see below.}

\item{strata}{(optional) A tab delimited file with 2 columns with header:
\code{INDIVIDUALS} and \code{STRATA}.
If a \code{strata} file is specified, the strata file will have
precedence over any grouping found input file (\code{data}). 
The \code{STRATA} column can be any hierarchical grouping.
Default: \code{strata = NULL}.}

\item{filename}{(optional) Name of the file prefix for
the matrix and the table written in the working directory.}
}
\value{
A list with 3 objects of class: table ($dlr.table), dist (a lower
diagonal matrix, $dlr.dist), data.frame (a mirrored matrix, $dlr.matrix).
}
\description{
The function computes Paetkau's et al. (1997) genotype likelihood
ratio distance (Dlr).
}
\references{
Paetkau D, Slade R, Burden M, Estoup A (2004) 
Genetic assignment methods for the direct, real-time estimation of 
migration rate: a simulation-based exploration of accuracy and power. 
Molecular Ecology, 13, 55-65.

Paetkau D, Waits LP, Clarkson PL, Craighead L, Strobeck C (1997)
 An empirical evaluation of genetic distance statistics using microsatellite
  data from bear (Ursidae) populations. Genetics, 147, 1943-1957.

Meirmans PG, Van Tienderen PH (2004) genotype and genodive: 
two programs for the analysis of genetic diversity of asexual organisms. 
Molecular Ecology Notes, 4, 792-794.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
