% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_WC84.R
\name{fst_WC84}
\alias{fst_WC84}
\title{A fast implementation of Weir and Cockerham (1984) Fst/Theta 
(overall and paiwise estimates)}
\arguments{
\item{data}{A file or object in the global environment containing at least 
these 4 columns: 
\code{INDIVIDUALS}, \code{POP_ID} (that refers to any grouping 
of individuals.), 
\code{MARKERS} and \code{GENOTYPES or GT} in a tidy format. During import, 
only those columns names will be kept.}

\item{sep}{(optional) A character string separating alleles. 
Default: \code{sep = NULL}.}

\item{pop.levels}{(optional, string) A character string with your populations 
ordered.
Default: \code{pop.levels = NULL}.}

\item{holdout.samples}{(optional) Samples that don't participate in the Fst 
computation (supplementary). Data frame with one column \code{INDIVIDUALS}.
Default: \code{holdout.samples = NULL}.}

\item{pairwise}{(logical, optional) With \code{pairwise = TRUE}, the 
pairwise WC84 Fst is calculated between populations. 
Default: \code{pairwise = NULL}.}

\item{ci}{(logical, optional) Compute bootstrapped confidence intervals. 
Default: \code{ci = NULL}.}

\item{iteration.ci}{(integer, optional) The number of iterations for 
the boostraps (resampling with replacement of markers). 
Default: \code{iteration.ci = 100}.}

\item{quantiles.ci}{(character, optional) 
The quantiles for the bootstrapped confidence intervals. 
Default: \code{quantiles.ci = c(0.025,0.975)}.}

\item{digits}{(optional, integer) The number of decimal places to be used in 
results.
Default: \code{digits = 4}.}

\item{parallel.core}{(optional) The number of core for parallel computation 
of pairwise Fst. 
If not selected \code{detectCores()-1} is used as default.}

\item{messages}{(logical, optional) Show messages during computations. 
Default: \code{messages = NULL}.}

\item{...}{other parameters passed to the function.}
}
\value{
With pairwise comparison computed, the function returns a list with 
9 objects: 
\code{$sigma.loc}: the variance components 
per locus (\code{lsiga}: among populations
\code{lsigb}: among individuals within populations
\code{lsigw}: within individuals),
\code{$fst.markers}: the fst by markers,
\code{$fst.ranked}: the fst ranked,
\code{$fst.overall}: the mean fst overall markers,
\code{$fis.markers}: the fis by markers,
\code{$fis.overall}: the mean fis overall markers,
\code{$pairwise.fst}: the pairwise fst in long format in a data frame,
\code{$pairwise.fst.matrix}: the pairwise fst in a upper triangle matrix.
\code{$pairwise.fst.ci.matrix}: the pairwise fst in the upper triangle 
matrix and the ci in the lower triangle matrix.
}
\description{
The function computes Weir and Cockerham (1984) 
Fst for diploid genomes.
Both overall and pairwise Fst can be estimated with confidence intervals 
based on 
bootstrap of markers (resampling with replacement). The function gives 
identical results (at the 9th decimal) to \code{\link[hierfstat]{genet.dist}} 
in \code{\link[hierfstat]{hierfstat}}, the Analysis of Molecular
Variance (AMOVA, Excoffier et al., 1992; Michalakis and Excoffier, 1996) and 
the Fst computed in \code{Calculate Distances} or \code{Paiwise Differentiation} 
options in 
\code{GenoDive} (Meirmans and Van Tienderen, 2004). \code{GenoDive} is still 
the fastest option, but for an R implementation, \code{\link[assigner]{fst_WC84}} 
is very fast and the computations takes advantage of \code{\link[dplyr]{dplyr}}, 
\code{tidyr}, \code{purrr} and 
\code{\link[parallel]{mclapply}} for parallel computing. 

This function is dedicated to Louis Bernatchez, 
in the hope that your students found the function fast and usefull.
}
\details{
Details for the sep argument:
This character is directly used in regular expressions using strigi. 
Some characters need to be preceeded by double backslashes \code{\\}. 
For instance, "/" works but "|" must be coded as "\\|".

From \code{GenoDive} manual:
'In general, rather than to test differentiation between all pairs of 
populations,
it is adviseable to perform an overall test of population differentiation, 
possibly using a hierarchical population structure, (see AMOVA)'

To compute an AMOVA, use \code{GenoDive} or \code{\link[mmod]{Phi_st_Meirmans}} 
in \code{\link[mmod]{mmod}}.
}
\examples{
\dontrun{
wombat.fst.pairwise <- fst_WC84(
data = "wombat.filtered.tidy.tsv", 
sep = "/", 
pop.levels = c("ATL", "MLE", "BIS", "PMO", "SOL", "TAS", "ECU"),
holdout.samples = NULL,
pairwise = TRUE,
ci = TRUE, 
iteration.ci = 10000, 
quantiles.ci = c(0.025,0.975),
parallel.core = 8,
messages = TRUE
)
To get the overall Fst estimate:
wombat.fst.pairwise$fst.overall
To get the pairwise Fst values with confidence intervals in a data frame:
wombat.fst.pairwise$pairwise.fst
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\references{
Excoffier L, Smouse PE, Quattro JM. 
Analysis of molecular variance inferred from metric distances among 
DNA haplotypes: application to human mitochondrial DNA restriction data. 
Genetics. 1992;131: 479-491.

Meirmans PG, Van Tienderen PH (2004) genotype and genodive: 
two programs for the analysis of genetic diversity of asexual organisms. 
Molecular Ecology Notes, 4, 792-794.

Michalakis Y, Excoffier L. 
A generic estimation of population 
subdivision using distances between alleles with special reference for 
microsatellite loci. 
Genetics. 1996;142: 1061-1064.

Weir BS, Cockerham CC (1984) Estimating F-Statistics for the
Analysis of Population Structure. 
Evolution, 38, 1358-1370.
}
\seealso{
\code{hierfstat} is available on 
CRAN \url{http://cran.r-project.org/web/packages/hierfstat/} and 
github \url{https://github.com/jgx65/hierfstat/}

\code{GenoDive} is available 
\url{http://www.bentleydrummer.nl/software/software/GenoDive.html}.

For Fisher's exact test and p-values per markers 
see \code{mmod} \code{\link[mmod]{diff_test}}.
}

