% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_WC84.R
\name{fst_WC84}
\alias{fst_WC84}
\title{A fast implementation of Weir and Cockerham (1984) Fst/Theta 
(overall and paiwise estimates)}
\usage{
fst_WC84(data, snprelate = FALSE, pop.levels = NULL, pop.labels = NULL,
  strata = NULL, holdout.samples = NULL, pairwise = FALSE, ci = FALSE,
  iteration.ci = 100, quantiles.ci = c(0.025, 0.975), subsample = NULL,
  iteration.subsample = 1, digits = 9,
  parallel.core = parallel::detectCores() - 1, verbose = FALSE, ...)
}
\arguments{
\item{data}{A file in the working directory or object in the global environment 
in wide or long (tidy) formats. To import, the function uses internally
\href{https://github.com/thierrygosselin/radiator}{radiator} 
\code{\link[radiator]{tidy_wide}}. See details for more info.

\emph{How to get a tidy data frame ?}
\href{https://github.com/thierrygosselin/radiator}{radiator} 
\code{\link[radiator]{tidy_genomic_data}} can transform 11 genomic data formats 
in a tidy data frame (VCF, PLINK, genind, genlight, gtypes, genepop,
stacks haplotype file, hierfstat, ...). 
You can also use this function to filter your dataset using
whitelist of markers, blacklist of individuals and genotypes.}

\item{snprelate}{(logical) Use \href{https://github.com/zhengxwen/SNPRelate}{SNPRelate}
to compute the Fst. Testing as shown an upward bias with \code{SNPRelate::snpgdsFst} function.
The values are 0.99 correlated with my built in codes and with \code{Hierfstat}.
But it's the fastest computation I've seen so far!
Default: \code{snprelate = FALSE}}

\item{pop.levels}{(optional, string) This refers to the levels in a factor. In this 
case, the id of the pop.
Use this argument to have the pop ordered your way instead of the default 
alphabetical or numerical order. e.g. \code{pop.levels = c("QUE", "ONT", "ALB")} 
instead of the default \code{pop.levels = c("ALB", "ONT", "QUE")}. 
Default: \code{pop.levels = NULL}.}

\item{pop.labels}{(optional, string) Use this argument to rename/relabel
your pop or combine your pop. e.g. To combine \code{"QUE"} and \code{"ONT"} 
into a new pop called \code{"NEW"}:
(1) First, define the levels for your pop with \code{pop.levels} argument: 
\code{pop.levels = c("QUE", "ONT", "ALB")}. 
(2) then, use \code{pop.labels} argument: 
\code{pop.levels = c("NEW", "NEW", "ALB")}.#' 
To rename \code{"QUE"} to \code{"TAS"}:
\code{pop.labels = c("TAS", "ONT", "ALB")}.
Default: \code{pop.labels = NULL}. If you find this too complicated, there is also the
\code{strata} argument that can do the same thing, see below.}

\item{strata}{(optional, data frame) A tab delimited file with 2 columns with header:
\code{INDIVIDUALS} and \code{STRATA}.
If a \code{strata} file is specified, the strata file will have
precedence over any grouping found input file (\code{data}). 
The \code{STRATA} column can be any hierarchical grouping.
Default: \code{strata = NULL}.}

\item{holdout.samples}{(optional, data frame) Samples that don't participate in the Fst 
computation (supplementary). Data frame with one column \code{INDIVIDUALS}.
Default: \code{holdout.samples = NULL}.}

\item{pairwise}{(optional, logical) With \code{pairwise = TRUE}, the 
pairwise WC84 Fst is calculated between populations. 
Default: \code{pairwise = FALSE}.}

\item{ci}{(optional, logical) Compute bootstrapped confidence intervals. 
Default: \code{ci = FALSE}.}

\item{iteration.ci}{(optional, integer) The number of iterations for 
the boostraps (resampling with replacement of markers). 
Default: \code{iteration.ci = 100}.}

\item{quantiles.ci}{(optional, double) 
The quantiles for the bootstrapped confidence intervals. 
Default: \code{quantiles.ci = c(0.025,0.975)}.}

\item{subsample}{(Integer or character) 
With \code{subsample = 36}, 36 individuals in each populations are chosen
randomly to represent the dataset. With \code{subsample = "min"}, the 
minimum number of individual/population found in the data is used automatically.
Default is no subsampling, \code{subsample = NULL}.}

\item{iteration.subsample}{(Integer) The number of iterations to repeat 
subsampling.
With \code{subsample = 20} and \code{iteration.subsample = 10},
20 individuals/populations will be randomly chosen 10 times.
Default: \code{iteration.subsample = 1}.}

\item{digits}{(optional, integer) The number of decimal places to be used in 
results.
Default: \code{digits = 9}.}

\item{parallel.core}{(optional, integer) The number of core for parallel computation 
of pairwise Fst.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{verbose}{(optional, logical) \code{verbose = TRUE} to be chatty 
during execution. 
Default: \code{verbose = FALSE}.}

\item{...}{other parameters passed to the function.}
}
\value{
The function returns a list with several objects.
When sumsample is selected the objects end with \code{.subsample}.
\itemize{
 \item \code{$subsampling.individuals}: the combinations of individuals and subsamples,
 \item \code{$sigma.loc}: the variance components per locus, with
      (\code{lsiga}: among populations, 
      \code{lsigb}: among individuals within populations,
      \code{lsigw}: within individuals)
 \item \code{$fst.markers}: the fst by markers,
 \item \code{$fst.ranked}: the fst ranked,
 \item \code{$fst.overall}: the mean fst overall markers and the number of markers 
 \item \code{$fis.markers}: the fis by markers,
 \item \code{$fis.overall}: the mean fis overall markers and the number of markers,
 \item \code{$fst.plot}: the histogram of the overall Fst per markers,
 \item \code{$pairwise.fst}: the pairwise fst in long/tidy data frame and the number of markers ,
 \item \code{$pairwise.fst.upper.matrix}: the pairwise fst in a upper triangle matrix,
 \item \code{$pairwise.fst.full.matrix}: the pairwise fst matrix (duplicated upper and lower triangle),
 \item \code{$pairwise.fst.ci.matrix}: matrix with pairwise fst in the upper triangle
 and the confidence intervals in the lower triangle.
 \item when subsample is selected \code{$pairwise.fst.subsample.mean} is a summary
 of all pairwise comparisons subsample. The mean is calculated accross summary
 statistics.
}
}
\description{
The function calculates Weir and Cockerham (1984) 
Fst for diploid genomes. Both overall and pairwise Fst can be estimated with 
confidence intervals based on bootstrap of markers (resampling with replacement). 
The function gives identical results \emph{at the 9th decimal} when tested 
against \code{\link[hierfstat]{genet.dist}} in \pkg{hierfstat}. Using the 
argument \code{snprelate = TRUE} will compute the Fst with 
\href{https://github.com/zhengxwen/SNPRelate}{SNPRelate}. This implementation
gives slightly upward bias values but provided the fastest computations I know,
but it doesn't compute confidence intervals, for now. 
For an R implementation, \code{\link{fst_WC84}} is very fast. 
The computations takes advantage of \pkg{dplyr}, \pkg{tidyr}, \pkg{purrr}, 
\pkg{data.table}, \pkg{parallel} and \pkg{SNPRelate}.
The impact of unbalanced design on estimates can be tested by using the 
subsample argument.

Note: Until the bias observed with SNPRelate is resolved the option will remain unavailable.

\emph{Special concerns for genome-wide estimate and filtering bias}

During computation, the function first starts by keeping only the polymorphic 
markers in common between the populations. Keep this in mind when filtering 
your markers to use this function characteristic strategically to get 
better genome-wide estimate. This is even more important when your project 
involves more than 2 populations that evolved more by neutral processes 
(e.g. genetic drift) than by natural selection (see the vignette for more details).
}
\details{
\strong{Input data:}
 
To discriminate the long from the wide format, 
the function \pkg{radiator} \code{\link[radiator]{tidy_wide}} searches 
for \code{MARKERS or LOCUS} in column names (TRUE = long format).
The data frame is tab delimitted.
\strong{Wide format:}
The wide format cannot store metadata info.
The wide format starts with these 2 id columns: 
\code{INDIVIDUALS}, \code{POP_ID} (that refers to any grouping of individuals), 
the remaining columns are the markers in separate columns storing genotypes.

\strong{Long/Tidy format:}
The long format is considered to be a tidy data frame and can store metadata info. 
(e.g. from a VCF see \pkg{radiator} \code{\link{tidy_genomic_data}}). A minimum of 4 columns
are required in the long format: \code{INDIVIDUALS}, \code{POP_ID}, 
\code{MARKERS or LOCUS} and \code{GENOTYPE or GT}. The rest are considered metata info.

\strong{2 genotypes formats are available:}
6 characters no separator: e.g. \code{001002 of 111333} (for heterozygote individual).
6 characters WITH separator: e.g. \code{001/002 of 111/333} (for heterozygote individual).
The separator can be any of these: \code{"/", ":", "_", "-", "."}.

\emph{How to get a tidy data frame ?}
\pkg{radiator} \code{\link{tidy_genomic_data}} can transform 6 genomic data formats 
in a tidy data frame.
}
\note{
\strong{Negative Fst} are technical artifact of the computation 
(see Roesti el al. 2012) and are automatically replaced with zero inside 
this function.

\strong{Why no p-values ?}

There is no null hypothesis testing with \emph{P}-values.
Confidence intervals provided with the \emph{F}-statistics 
enables more reliable conclusions about the biological trends in the data.
}
\examples{
\dontrun{
wombat.fst.pairwise <- fst_WC84(
data = "wombat.filtered.tidy.tsv", 
sep = "/", 
pop.levels = c("ATL", "MLE", "BIS", "PMO", "SOL", "TAS", "ECU"),
holdout.samples = NULL,
pairwise = TRUE,
ci = TRUE, 
iteration.ci = 10000, 
quantiles.ci = c(0.025,0.975),
parallel.core = 8,
verbose = TRUE
)
To get the overall Fst estimate:
wombat.fst.pairwise$fst.overall
To get the Fst plot:
wombat.fst.pairwise$fst.plots
To get the pairwise Fst values with confidence intervals in a data frame:
wombat.fst.pairwise$pairwise.fst
To get the pairwise fst and ci matrix in a data frame:
# rename, data frame, put rownames in column
pairwise.fst.ci.df <- data.frame(pairwise.fst.ci.matrix) \%>\% add_rownames("POP")
}
}
\references{
Excoffier L, Smouse PE, Quattro JM. 
Analysis of molecular variance inferred from metric distances among 
DNA haplotypes: application to human mitochondrial DNA restriction data. 
Genetics. 1992;131: 479-491.

Meirmans PG, Van Tienderen PH (2004) genotype and genodive: 
two programs for the analysis of genetic diversity of asexual organisms. 
Molecular Ecology Notes, 4, 792-794.

Michalakis Y, Excoffier L. 
A generic estimation of population 
subdivision using distances between alleles with special reference for 
microsatellite loci. 
Genetics. 1996;142: 1061-1064.

Weir BS, Cockerham CC (1984) Estimating F-Statistics for the
Analysis of Population Structure. 
Evolution, 38, 1358-1370.

Roesti M, Salzburger W, Berner D. (2012)
Uninformative polymorphisms bias genome scans for signatures of selection. 
BMC Evol Biol., 12:94. doi:10.1111/j.1365-294X.2012.05509.x

Zheng X, Levine D, Shen J, Gogarten SM, Laurie C, Weir BS.
A high-performance computing toolset for relatedness and principal component
analysis of SNP data. Bioinformatics. 2012;28: 3326-3328.
doi:10.1093/bioinformatics/bts606
}
\seealso{
From \href{http://www.bentleydrummer.nl/software/software/GenoDive.html}{GenoDive} manual:
\emph{'In general, rather than to test differentiation between all pairs of 
populations,
it is adviseable to perform an overall test of population differentiation, 
possibly using a hierarchical population structure, (see AMOVA)'}

To compute an AMOVA, use \href{http://www.bentleydrummer.nl/software/software/GenoDive.html}{GenoDive}
or \code{\link[mmod]{Phi_st_Meirmans}} 
in \code{\link[mmod]{mmod}}.

\code{hierfstat} is available on 
CRAN \url{http://cran.r-project.org/web/packages/hierfstat/} and 
github \url{https://github.com/jgx65/hierfstat/}

Link for \href{http://www.bentleydrummer.nl/software/software/GenoDive.html}{GenoDive}

For Fisher's exact test and p-values per markers 
see \code{mmod} \code{\link[mmod]{diff_test}}.

\code{\link[radiator]{tidy_genomic_data}} to transform numerous genomic data 
format in tidy data frames.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
