% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_ngs.R
\name{generate_holdhout}
\alias{generate_holdhout}
\title{generate_holdhout}
\usage{
generate_holdhout(
  thl,
  strata,
  iteration.method,
  random.seed = NULL,
  path.folder = NULL
)
}
\description{
Generate holdhout samples based on thl and iterations.list
}
\keyword{internal}
