% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_ngs.R
\name{generate_subsamples}
\alias{generate_subsamples}
\title{generate_subsamples}
\usage{
generate_subsamples(
  subsample,
  iteration.subsample,
  strata,
  random.seed = NULL,
  path.folder = NULL
)
}
\description{
generate_subsamples
}
\keyword{internal}
