% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_mixture.R, R/assignment_ngs.R
\name{marker_selection}
\alias{marker_selection}
\title{marker_selection}
\usage{
marker_selection(iteration.method, m = NULL, random.seed = NULL,
  unique.markers = NULL)

marker_selection(iteration.method, m = NULL, random.seed = NULL,
  unique.markers = NULL)
}
\description{
Random selection of marker function + iteration.method

Random selection of marker function + iteration.method
}
\keyword{internal}

