% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlr.R
\name{plot_dlr}
\alias{plot_dlr}
\title{Assignment plot of genotype likelihood (Dlr).}
\usage{
plot_dlr(data, pop.pairwise, dlr = NULL, x.dlr = NULL, y.dlr = NULL,
  fst = NULL, x.fst = NULL, y.fst = NULL, filename = NULL,
  plot.width = NULL, plot.height = NULL, plot.dpi = NULL)
}
\arguments{
\item{data}{The assigment object from
\code{\link[assigner]{dlr}} output.}

\item{pop.pairwise}{List of pairwise populations to generate the Dlr plot.}

\item{dlr}{(optional) Character string with Dlr value.}

\item{x.dlr}{(optional) Position to the x-axis of the Dlr value.}

\item{y.dlr}{(optional) Position to the y-axis of the Dlr value.}

\item{fst}{(optional) Character string with Fst value.}

\item{x.fst}{(optional) Position to the x-axis of the Fst value.}

\item{y.fst}{(optional) Position to the y-axis of the Fst value.}

\item{filename}{(optional) Name of the figure written
in the working directory.}

\item{plot.width}{(optional) Width in cm of the figure.}

\item{plot.height}{(optional) height in cm of the figure.}

\item{plot.dpi}{(optional) Number of dpi for the figure (e.g 600).}
}
\value{
A list with the assignment table and the assignment plot.
}
\description{
The function generate a figure similar to Paetkau's et al. (2004)
Fig 6.
}
\references{
Paetkau D, Slade R, Burden M, Estoup A (2004)
Genetic assignment methods for the direct, real-time estimation of migration
rate: a simulation-based exploration of accuracy and power.
Molecular Ecology, 13, 55-65.

Meirmans PG, Van Tienderen PH (2004) genotype and genodive:
two programs for the analysis of genetic diversity of asexual organisms.
Molecular Ecology Notes, 4, 792-794.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\keyword{internal}
