% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_mixture.R, R/subsampling_data.R
\name{subsampling_data}
\alias{subsampling_data}
\alias{subsampling_data}
\title{subsampling data}
\usage{
subsampling_data(iteration.subsample = 1, ind.pop.df = NULL,
  subsample = NULL, random.seed = NULL)

subsampling_data(iteration.subsample = 1, ind.pop.df = NULL,
  subsample = NULL, random.seed = NULL)
}
\description{
subsampling data

subsampling data
}
\keyword{internal}
